new: limitAcceleration + extra
In process of creating system to cap the possible jump in velocities that would result in undesirable accelerations. This is just a rough draft of the method and I will make it less convoluted later. Also added MAX_ACCELERATION to constants class. In process of creating helper method in drive class to turn MAX_ACCELERATION into a velocity vector over the period that swerveDrive would be called.