rework math in VisionManager to make it make more sense
The code should still produce the same outputs, but it should make more sense when trying to read it.

Basically when calculating the angle we were calculating the angle from the goal to the robot when we wanted the angle from the robot to the goal. Because we shoot backwards this worked, but it doesn't make immediate sense when reading the code.

To fix this we we multiplied the translation by negative one (which will give us a vector from the robot to the goal). Since we still need to shoot backwards we then added a method to rotate this translation by 180 so that our robot still aims correctly.